Classes to run cooperative multitasking
This file contains classes to run cooperatively scheduled tasks in a multitasking system. Tasks are created as generators, functions which have infinite loops and call yield at least once in the loop. References to all the tasks to be run in the system are kept in a list maintained by class CoTaskList; the  system scheduler then runs the tasks' run() methods according to a chosen scheduling algorithm such as round-robin or highest-priority-first.